requirements.txt
Flask==2.3.3
openai-whisper==20231117
openai==1.51.0
Werkzeug==2.3.7
gunicorn==21.2.0